requirements.txt
apturl==0.5.2
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
contourpy==1.2.0
cryptography==43.0.1
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.12.1
daphne==4.1.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
fonttools==4.44.3
freetype-py==2.3.0
future==0.18.2
h11==0.13.0
html5lib==1.1
httplib2==0.20.2
huepy==1.2.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
livereload==2.6.3
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown2==2.4.13
MarkupSafe==2.0.1
matplotlib==3.8.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
oscrypto==1.3.0
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
pypdf==3.17.0
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4.2
razorpay==1.4.1
redis==5.0.8
reportlab==4.0.7
requests==2.25.1
rlPyCairo==0.3.0
SecretStorage==3.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
systemd-python==234
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
ttystatus==0.38
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.8.0
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.2
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
virtualenv==20.13.0+ds
vite==1.5.2
wadllib==1.3.6
webencodings==0.5.1
wsproto==1.0.0
xdg==5
xhtml2pdf==0.2.13
xkit==0.0.0
zipp==1.0.0
zope.interface==7.0.3
